put create and destroy graph div routine in jasmine tests assets/
