+++ test/jasmine/assets/create_graph_div.js
@@ -1,0 +1,8 @@
+'use strict';
+
+module.exports = function createGraphDiv() {
+    var gd = document.createElement('div');
+    gd.id = 'graph';
+    document.body.appendChild(gd);
+    return gd;
+};
