+++ test/jasmine/assets/destroy_graph_div.js
@@ -1,0 +1,6 @@
+'use strict';
+
+module.exports = function destroyGraphDiv() {
+    var gd = document.getElementById('graph');
+    document.body.removeChild(gd);
+};
