+++ test/jasmine/tests/gl_plot_interact_test.js
@@ -2,6 +2,9 @@ var d3 = require('d3');
 
 var Plotly = require('@src/index');
 
+var createGraphDiv = require('../assets/create_graph_div');
+var destroyGraphDiv = require('../assets/destroy_graph_div');
+
 /*
  * WebGL interaction test cases fail on the CircleCI
  * most likely due to a WebGL/driver issue
@@ -11,18 +14,6 @@ var Plotly = require('@src/index');
 
 describe('Test plot structure', function () {
     'use strict';
-
-    function createGraphDiv() {
-        var gd = document.createElement('div');
-        gd.id = 'graph';
-        document.body.appendChild(gd);
-        return gd;
-    }
-
-    function destroyGraphDiv() {
-        var gd = document.getElementById('graph');
-        document.body.removeChild(gd);
-    }
 
     afterEach(destroyGraphDiv);
